Booking api tests missing email testing.

+++ Buchungssystem_Lastenkarle/api/algorithm.py
@@ -90,31 +90,38 @@ def merge_availabilities_algorithm(booking):
                                                        until_date=end_booking.isoformat(),
                                                        bike_id=booking.bike.pk,
                                                        store_id=booking.bike.store.pk)
-    left_side = Availability.objects.filter(until_date=(begin_booking - timedelta(days=1)).isoformat(),
-                                            bike_id=booking.bike.pk,
-                                            store_id=booking.bike.store.pk)
-
     new_from_date = booking.begin.isoformat()
     new_until_date = booking.end.isoformat()
-    if left_side.filter(availability_status=Availability_Status.objects.get(availability_status='Booked')).exists():
-        new_from_date = begin_booking
-    else:
-        if left_side.filter(
-                availability_status=Availability_Status.objects.get(availability_status='Available')).exists():
-            new_from_date = left_side.first().from_date
-            left_side.delete()
-
-    right_side = Availability.objects.filter(from_date=(end_booking + timedelta(days=1)).isoformat(),
-                                             bike_id=booking.bike.pk,
-                                             store_id=booking.bike.store.pk)
-
-    if right_side.filter(availability_status=Availability_Status.objects.get(availability_status='Booked')).exists():
+    if begin_booking.isoformat() == date(1000, 1, 1).isoformat():
         new_until_date = end_booking
     else:
+        left_side = Availability.objects.filter(until_date=(begin_booking - timedelta(days=1)).isoformat(),
+                                                bike_id=booking.bike.pk,
+                                                store_id=booking.bike.store.pk)
+
+        if left_side.filter(availability_status=Availability_Status.objects.get(availability_status='Booked')).exists():
+            new_from_date = begin_booking
+        else:
+            if left_side.filter(
+                    availability_status=Availability_Status.objects.get(availability_status='Available')).exists():
+                new_from_date = left_side.first().from_date
+                left_side.delete()
+
+    if end_booking.isoformat() == date(5000, 1, 1).isoformat():
+        new_until_date = end_booking
+    else:
+        right_side = Availability.objects.filter(from_date=(end_booking + timedelta(days=1)).isoformat(),
+                                                 bike_id=booking.bike.pk,
+                                                 store_id=booking.bike.store.pk)
+
         if right_side.filter(
-                availability_status=Availability_Status.objects.get(availability_status='Available')).exists():
-            new_until_date = right_side.first().until_date
-            right_side.delete()
+                availability_status=Availability_Status.objects.get(availability_status='Booked')).exists():
+            new_until_date = end_booking
+        else:
+            if right_side.filter(
+                    availability_status=Availability_Status.objects.get(availability_status='Available')).exists():
+                new_until_date = right_side.first().until_date
+                right_side.delete()
 
     merged_availability = Availability.objects.create(from_date=new_from_date,
                                                       until_date=new_until_date,
